Add jwtCheck to verify user Authorization
